Implement APSTUD-4874 Add ability to toggle HTML Tidy rules on and off - Place the problems into categories in UI, make the categories expandable/collapsable. - Add a Severity enum to IProblem that encapsulates the integer value used for markers, adds "ignore" as an options, holds UI label - Add test cases for maniuplating severities of problem types - handle restoring problem severity prefs - handle looking up severities when generating problems in HTML Tidy - Use pref initializer to set id/name attribute mismatch problem as an "error" by default for HTMLTidy - Add ability to store integer values in prefs of participant

+++ tests/com.aptana.editor.css.tests/src/com/aptana/editor/css/validator/CSSValidatorTest.java
@@ -101,7 +101,7 @@ public class CSSValidatorTest extends AbstractValidatorTestCase
 
 		// makes sure it's just a warning unrelated to the src
 		IProblem item = items.get(0);
-		assertEquals(IMarker.SEVERITY_WARNING, item.getSeverity());
+		assertEquals(IMarker.SEVERITY_WARNING, item.getSeverity().intValue());
 	}
 
 	public void testCSS3AtRule() throws CoreException
